+++ src/hooks/use-infinite-scroll/__test__/index.test.tsx
@@ -1,6 +1,6 @@
 import { MutableRefObject, useRef } from 'react';
-import { act, fireEvent, render } from '@testing-library/react';
-import { useIsTouchDevice, useInfiniteScroll, useOutsideClick } from '../index';
+import { act, render } from '@testing-library/react';
+import { useInfiniteScroll } from '..';
 
 describe('useInfiniteScroll()', () => {
   const TestComponent = ({
@@ -93,91 +93,5 @@ describe('useInfiniteScroll()', () => {
     expect((listElement as unknown as HTMLUListElement).removeEventListener).toHaveBeenCalledTimes(
       1,
     );
-  });
-});
-
-describe('useIsTouchDevice', () => {
-  const TestComponent = () => {
-    const touch = useIsTouchDevice();
-
-    return <>{touch && <span>Visible on touch</span>}</>;
-  };
-
-  it('should works with touch', () => {
-    // simulate touch device
-    window.ontouchstart = undefined;
-
-    const { container } = render(<TestComponent />);
-
-    expect(container.querySelector('span')?.textContent).toBe('Visible on touch');
-  });
-});
-
-describe('useOutsideClick', () => {
-  const TestComponent = ({ callback }: { callback: () => void }) => {
-    const ref = useRef<HTMLDivElement>(null);
-
-    useOutsideClick(ref, callback);
-
-    return <div data-testid='target' ref={ref} />;
-  };
-
-  it('should works', () => {
-    const spy = jest.fn();
-
-    render(<TestComponent callback={spy} />);
-
-    expect(spy).toHaveBeenCalledTimes(0);
-
-    act(() => {
-      document.documentElement.dispatchEvent(new MouseEvent('click'));
-    });
-
-    expect(spy).toHaveBeenCalledTimes(1);
-  });
-
-  it('should handle array of refs', () => {
-    const spy = jest.fn();
-
-    const OtherTestComponent = ({ callback }: { callback: () => void }) => {
-      const refA = useRef<HTMLDivElement>(null);
-      const refB = useRef<HTMLDivElement>(null);
-
-      useOutsideClick([refA, refB], callback);
-
-      return (
-        <>
-          <div ref={refA} />
-          <div ref={refB} />
-        </>
-      );
-    };
-
-    render(<OtherTestComponent callback={spy} />);
-
-    expect(spy).toHaveBeenCalledTimes(0);
-
-    act(() => {
-      document.documentElement.dispatchEvent(new MouseEvent('click'));
-    });
-
-    expect(spy).toHaveBeenCalledTimes(1);
-  });
-
-  it('should not call callback', () => {
-    const spy = jest.fn();
-
-    const { getByTestId } = render(
-      <>
-        <span data-testid='sibling'></span>
-        <TestComponent callback={spy} />
-      </>,
-    );
-
-    expect(spy).toHaveBeenCalledTimes(0);
-    fireEvent.click(getByTestId('target'));
-    expect(spy).toHaveBeenCalledTimes(0);
-    fireEvent.click(getByTestId('sibling'));
-    expect(spy).toHaveBeenCalledTimes(1);
   });
 });
