+++ src/hooks/use-isomorphic-layout-effect/__test__/index.web.test.ts
@@ -2,7 +2,7 @@
  * @jest-environment jsdom
  */
 import { useLayoutEffect } from 'react';
-import { useIsomorphicLayoutEffect } from '../..';
+import { useIsomorphicLayoutEffect } from '..';
 
 describe('useIsomorphicLayoutEffect', () => {
   it('should work as regular effect in NodeJS environment', () => {
